Improve performance of median pooling
This patch improves the performance of `median_pool()`.
For `uint8` arrays, median pooling is now between
1.7x and 7x faster, if the kernel size is 3 or 5
or if the kernel size is 7, 9, 11, 13 and the image
size is <=32x32.
In both cases the kernel also has to be symmetric.
In the case of a kernel size of 3, the
performance improvement is most significant for
larger images. In the case of 5, it is fairly
even over all kernel sizes. In the case of
>=7, it is more significant for smaller images.